2313: cleanup
