+++ native/src/components/SprungbrettListItem.tsx
@@ -17,11 +17,7 @@ class SprungbrettListItem extends PureComponent<SprungbrettListItemProps> {
   render(): ReactElement {
     const { language, job, openJobInBrowser } = this.props
     return (
-      <ListItem
-        thumbnail={<SprungbrettIcon width={24} height={24} />}
-        title={job.title}
-        navigateTo={openJobInBrowser}
-        language={language}>
+      <ListItem thumbnail={<SprungbrettIcon />} title={job.title} navigateTo={openJobInBrowser} language={language}>
         <Text>{job.location}</Text>
       </ListItem>
     )
